Redo active logic
Using MutationObserver is not necessary since the only place where this isn't handled was when toggling active for consent-overlay.div. I do however think that using this is a good practice since the overhead shouldn't be massive and it should handle all cases. It also at least somewhat improves readability.